requirements.txt
pymongo
requests
Flask
Flask-Cors
python-dotenv
secure-smtplib
razorpay==1.4.1
requests==2.31.0
openai
pillow